requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.0.6
django-filter==1.1.0
django-rest-framework-mongoengine==3.3.1
djangorestframework==3.8.2
freeze==1.0.10
idna==2.7
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
mongoengine==0.15.0
mysqlclient==1.3.12
pymongo==3.6.1
pytz==2018.4
requests==2.19.0
six==1.11.0
uritemplate==3.0.0
urllib3==1.23